refactor: utils are now separated and simpler to understand
