--- src/utils.ts
@@ -1,119 +1,0 @@
-import { GIVEAWAYS_REPO, SELECTORS } from "./constants";
-import { logNetworkError, logDomError } from "./loggers";
-
-// Cache for giveaways to avoid multiple requests
-let giveawaysCache: Map<string, string[]> | null = null;
-
-/**
- * Waits for a given time in milliseconds
- * @param time_ms time in milliseconds to sleep
- */
-export async function sleep(time_ms: number): Promise<void> {
-  await new Promise((resolve) => setTimeout(resolve, time_ms));
-}
-
-/**
- * Loads the giveaways from the server
- * @returns a map of regions and their respective giveaway names
- * @throws Error if the request fails
- */
-export async function giveaways(): Promise<Map<string, string[]>> {
-  if (giveawaysCache) {
-    return giveawaysCache;
-  }
-
-  try {
-    const response = await GM.xmlHttpRequest({
-      method: "GET",
-      url: GIVEAWAYS_REPO,
-      headers: { "Content-Type": "application/json" },
-    });
-
-    if (response.status !== 200) {
-      throw new Error(`Failed to load giveaways: ${response.statusText}`);
-    }
-
-    // Tampermonkey has `response.responseText` but Greasemonkey has `response.response`
-    const text = response.responseText ?? response.response;
-
-    const obj = JSON.parse(text);
-    // Convert the object to a Map
-    // this is necessary as a normal object would contain more properties
-    giveawaysCache = new Map(Object.entries(obj));
-    return giveawaysCache;
-  } catch (error) {
-    logNetworkError(error, "giveaways");
-    throw new Error("Unable to load giveaways");
-  }
-}
-
-/**
- * @param doc Document to check
- * @returns the button to click to validate the participation or null if not found
- */
-export function getValidationButton(doc: Document): HTMLButtonElement | null {
-  return doc.querySelector(SELECTORS.participateButton);
-}
-
-/**
- * Tries to get the element with the given selector,
- * if not found waits for the dom to update within the time limit
- * to retrieve the element
- * @param doc Document to check
- * @param selector the selector of the element to retrieve
- * @param timeout max time to wait for the element to be found
- * @returns the element if found, null otherwise
- */
-export function waitForElement<T extends Element>(
-  doc: Document,
-  selector: string,
-  timeout: number = 5000
-): Promise<T | null> {
-  return new Promise((resolve) => {
-    // Check if there is at least one element matching the selector
-    const element = doc.querySelector<T>(selector);
-    if (element !== null) {
-      resolve(element);
-      return;
-    }
-
-    // use a mutation observer to wait for the element to be added
-    // to the DOM
-    const observer = new MutationObserver(() => {
-      const element = doc.querySelector<T>(selector);
-      if (element !== null) {
-        observer.disconnect();
-        resolve(element);
-      }
-    });
-
-    // Observe changes in the DOM, including direct children (childList)
-    // and all descendant nodes (subtree) to detect the addition of the target element.
-    observer.observe(doc.body, { childList: true, subtree: true });
-
-    // If the element is not found within the timeout disconnect the observer
-    // and re-query for the last time in case it appeared just before the timeout
-    setTimeout(() => {
-      observer.disconnect();
-      // Re-query for the element in case it appeared before the timeout
-      const latestElement = doc.querySelector<T>(selector);
-      if (!latestElement) {
-        logDomError(`Element '${selector}' not found after ${timeout}ms`, "waitForElement");
-      }
-      resolve(latestElement);
-    }, timeout);
-  });
-}
-
-/**
- * Calculates the total number of giveaways
- * @param giveaways Map of regions and their respective giveaway names
- * @returns total number of giveaways
- */
-export function calculateTotal(giveaways: Map<string, string[]>): number {
-  let total = 0;
-  for (const names of giveaways.values()) {
-    total += names.length;
-  }
-  return total;
-}
